requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Twilio
twilio==8.10.0

# OpenAI (for LLM, STT, TTS)
openai==1.3.7

# File handling
pandas==2.1.3
python-docx==1.1.0
PyPDF2==3.0.1

# Cloudinary
cloudinary==1.41.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Google OAuth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

